Move `update_environment` from `run` to the `project` namespace
